Shopme/ShopmeWebParent/ShopmeBackEnd/src/main/java/com/shopme/admin/repository/UserRepository.java
package com.shopme.admin.repository;

import com.shopme.common.entity.User;
import org.springframework.data.domain.Page;
import org.springframework.data.domain.Pageable;
import org.springframework.data.jpa.repository.JpaRepository;
import org.springframework.data.jpa.repository.Query;
import org.springframework.data.repository.query.Param;

public interface UserRepository extends JpaRepository<User,Integer> {
    User findByEmail(String email);

    //By using Custom for finding unique customer

    @Query("SELECT u FROM User u WHERE u.email = :email")
    public User getUserByEmail(@Param("email") String email);


    @Query("SELECT u FROM User u WHERE CONCAT(u.id,' ',u.email,' ',u.firstName,' ',u.lastName) Like %?1% ")
    public Page<User>  findAll(String keyword, Pageable pageable);


}
